Maturity Level 2

+++ sajadv-backend/src/main/java/com/softplan/sajadv/user/service/UserService.java
@@ -8,14 +8,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.softplan.sajadv.user.exception.UserNotFoundException;
-import com.softplan.sajadv.user.exception.CPFRegistrationException;
 import com.softplan.sajadv.user.repository.UserRepository;
 
 @Service
 public class UserService {
 
-	@Autowired
-	private UserRepository userRepository;
+	private final UserRepository userRepository;
+
+	public UserService(UserRepository userRepository) {
+		this.userRepository = userRepository;
+	}
 
 	public List<User> getAll() {
 		return userRepository.findAll();
@@ -29,29 +31,21 @@ public class UserService {
 		return userRepository.findByCpf(cpf);
 	}
 
-	public User save(User user) {
-		
-		if (user.getId() == null) {
-			Optional<User> findByCpf = Optional.ofNullable(this.userRepository.findByCpf(user.getCpf())
-					.orElseThrow(() -> new UserNotFoundException(user.getId())));
-
-			if (findByCpf.isPresent()) {
-				throw new CPFRegistrationException("Cpf já cadastrado na base");
-			}
-		}
-		return userRepository.save(user);
+	public Optional<User> save(User user) {
+		return Optional.of(userRepository.save(user));
 	}
 
-	public User update(User user) {
-
-		if (user.getId() == null) {
+	public Optional<User> updateUser(User user) {
+		if (!userRepository.existsById(user.getId())) {
 			throw new UserNotFoundException(user.getId());
 		}
-
-		return userRepository.save(user);
+		return Optional.of(userRepository.save(user));
 	}
 
 	public void delete(Long id) {
+		if (!userRepository.existsById(id)) {
+			throw new UserNotFoundException(id);
+		}
 		userRepository.deleteById(id);
 	}
 }
